Factor out plugin config; use new config
