+++ packages/modular-scripts/react-scripts/config/webpack.config.js
@@ -3,29 +3,23 @@
 const chalk = require('chalk');
 const fs = require('fs');
 const path = require('path');
-const webpack = require('webpack');
-const resolve = require('resolve');
+
+const { merge } = require('webpack-merge');
+
 const builtinModules = require('builtin-modules');
 const PnpWebpackPlugin = require('pnp-webpack-plugin');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
-const CaseSensitivePathsPlugin = require('case-sensitive-paths-webpack-plugin');
-const InlineChunkHtmlPlugin = require('../../react-dev-utils/InlineChunkHtmlPlugin');
+
 const TerserPlugin = require('terser-webpack-plugin');
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const CssMinimizerPlugin = require('css-minimizer-webpack-plugin');
-const { WebpackManifestPlugin } = require('webpack-manifest-plugin');
 const logger = require('../../react-dev-utils/logger');
-const InterpolateHtmlPlugin = require('../../react-dev-utils/InterpolateHtmlPlugin');
-const WatchMissingNodeModulesPlugin = require('../../react-dev-utils/WatchMissingNodeModulesPlugin');
+const postcssNormalize = require('postcss-normalize');
+
 const ModuleScopePlugin = require('../../react-dev-utils/ModuleScopePlugin');
 const getCSSModuleLocalIdent = require('../../react-dev-utils/getCSSModuleLocalIdent');
 const paths = require('./paths');
 const modules = require('./modules');
-const getClientEnvironment = require('./env');
-const ModuleNotFoundPlugin = require('../../react-dev-utils/ModuleNotFoundPlugin');
-const ForkTsCheckerWebpackPlugin = require('fork-ts-checker-webpack-plugin');
-const postcssNormalize = require('postcss-normalize');
-const isCI = require('is-ci');
+const createPluginConfig = require('./pluginConfig');
 
 const isApp = process.env.MODULAR_IS_APP === 'true';
 const isEsmView = !isApp;
@@ -58,10 +52,6 @@ const reactRefreshOverlayEntry = require.resolve(
   '../../react-dev-utils/refreshOverlayInterop',
 );
 
-// Some apps do not need the benefits of saving a web request, so not inlining the chunk
-// makes for a smoother build process.
-const shouldInlineRuntimeChunk = process.env.INLINE_RUNTIME_CHUNK !== 'false';
-
 const imageInlineSizeLimit = parseInt(
   process.env.IMAGE_INLINE_SIZE_LIMIT || '10000',
 );
@@ -83,11 +73,11 @@ module.exports = function (webpackEnv) {
   const isEsmViewDevelopment = isEsmView & isEnvDevelopment;
 
   // This is needed if we're serving a ESM view in development node, since it won't be defined in the view dependencies.
-  if (isEsmViewDevelopment && externalDependencies?.react) {
-    externalDependencies['react-dom'] = externalDependencies?.react;
+  if (isEsmViewDevelopment && externalDependencies.react) {
+    externalDependencies['react-dom'] = externalDependencies.react;
   }
-  if (isEsmViewDevelopment && externalResolutions?.react) {
-    externalResolutions['react-dom'] = externalResolutions?.react;
+  if (isEsmViewDevelopment && externalResolutions.react) {
+    externalResolutions['react-dom'] = externalResolutions.react;
   }
 
   // Create a map of external dependencies if we're building a ESM view
@@ -103,12 +93,6 @@ module.exports = function (webpackEnv) {
   // passed into alias object. Uses a flag if passed into the build command
   const isEnvProductionProfile =
     isEnvProduction && process.argv.includes('--profile');
-
-  // We will provide `paths.publicUrlOrPath` to our app
-  // as %PUBLIC_URL% in `index.html` and `process.env.PUBLIC_URL` in JavaScript.
-  // Omit trailing slash as %PUBLIC_URL%/xyz looks better than %PUBLIC_URL%xyz.
-  // Get environment variables to inject into our app.
-  const env = getClientEnvironment(paths.publicUrlOrPath.slice(0, -1));
 
   // common function to get style loaders
   const getStyleLoaders = (cssOptions, preProcessor, includeEsmLoader) => {
@@ -183,146 +167,98 @@ module.exports = function (webpackEnv) {
     return loaders;
   };
 
-  const webpackConfig = {
-    externals: isApp
-      ? undefined
-      : function ({ request }, callback) {
-          const parsedModule = parsePackageName(request);
-
-          // If the module is absolute and it is in the import map, we want to externalise it
-          if (
-            parsedModule &&
-            parsedModule.dependencyName &&
-            dependencyMap[parsedModule.dependencyName] &&
-            // If this is an absolute export of css we need to deal with it in the loader
-            !request.endsWith('.css')
-          ) {
-            const { dependencyName, submodule } = parsedModule;
-
-            const toRewrite = `${dependencyMap[dependencyName]}${
-              submodule ? `/${submodule}` : ''
-            }`;
-
-            return callback(null, toRewrite);
-          }
-          // Otherwise we just want to bundle it
-          return callback();
-        },
-
-    externalsType: isApp ? undefined : 'module',
+  const appConfig = {
+    // TODO: remove me
     experiments: {
-      outputModule: isApp ? undefined : true,
+      outputModule: undefined,
     },
+    externals: undefined,
+    externalsType: undefined,
+    output: {
+      library: undefined,
+      module: undefined,
+      path: undefined,
+    },
+    // TODO: end remove
+    optimization: {
+      // Automatically split vendor and commons
+      splitChunks: { chunks: 'all' },
+      // Keep the runtime chunk separated to enable long term caching
+      // https://twitter.com/wSokra/status/969679223278505985
+      // https://github.com/facebook/create-react-app/issues/5358
+      runtimeChunk: {
+        name: (entrypoint) => `runtime-${entrypoint.name}`,
+      },
+    },
+  };
+
+  const esmViewConfig = {
+    externals: rewriteExternals,
+    externalsType: 'module',
+    experiments: { outputModule: true },
+    output: {
+      module: true,
+      library: { type: 'module' },
+    },
+  };
+
+  const productionConfig = {
+    mode: 'production',
+    // Stop compilation early in production
+    bail: true,
+    devtool: shouldUseSourceMap ? 'source-map' : false,
+    output: {
+      // The build folder.
+      path: paths.appBuild,
+      pathinfo: false,
+      // There will be one main bundle, and one file per asynchronous chunk.
+      filename: 'static/js/[name].[contenthash:8].js',
+      // There are also additional JS chunk files if you use code splitting.
+      // Please remember that Webpack 5, unlike Webpack 4, controls "splitChunks" via fileName, not chunkFilename - https://stackoverflow.com/questions/66077740/webpack-5-output-chunkfilename-not-working
+      chunkFilename: 'static/js/[name].[contenthash:8].chunk.js',
+      // Point sourcemap entries to original disk location (format as URL on Windows)
+      devtoolModuleFilenameTemplate: (info) =>
+        path
+          .relative(paths.appSrc, info.absoluteResourcePath)
+          .replace(/\\/g, '/'),
+    },
+    optimization: {
+      minimize: true,
+    },
+  };
+
+  const developementConfig = {
+    mode: 'development',
+    bail: false,
+    devtool: 'cheap-module-source-map',
+    output: {
+      // Add /* filename */ comments to generated require()s in the output.
+      pathinfo: true,
+      // In development, it does not produce real files.
+      filename: 'static/js/[name].js',
+      // Please remember that Webpack 5, unlike Webpack 4, controls "splitChunks" via fileName, not chunkFilename - https://stackoverflow.com/questions/66077740/webpack-5-output-chunkfilename-not-working
+      chunkFilename: 'static/js/[name].chunk.js',
+      // Point sourcemap entries to original disk location (format as URL on Windows)
+      devtoolModuleFilenameTemplate: (info) =>
+        path.resolve(info.absoluteResourcePath).replace(/\\/g, '/'),
+    },
+    optimization: {
+      minimize: false,
+    },
+  };
+
+  const baseConfig = {
     // Workaround for this bug: https://stackoverflow.com/questions/53905253/cant-set-up-the-hmr-stuck-with-waiting-for-update-signal-from-wds-in-cons
     target: 'web',
-    mode: isEnvProduction ? 'production' : isEnvDevelopment && 'development',
-    // Stop compilation early in production
-    bail: isEnvProduction,
-    devtool: isEnvProduction
-      ? shouldUseSourceMap
-        ? 'source-map'
-        : false
-      : isEnvDevelopment && 'cheap-module-source-map',
     // These are the "entry points" to our application.
     // This means they will be the "root" imports that are included in JS bundle.
     // We bundle a virtual file to trampoline the ESM view as an entry point if we're starting it (ESM views have no ReactDOM.render)
-    entry: isEsmViewDevelopment ? getVirtualTrampoline() : paths.appIndexJs,
+    entry: isEsmViewDevelopment ? getVirtualTrampoline() : paths.appIndexJs, // REWRITE
     output: {
-      module: isApp ? undefined : true,
-      library: isApp
-        ? undefined
-        : {
-            type: 'module',
-          },
-      // The build folder.
-      path: isEnvProduction ? paths.appBuild : undefined,
-      // Add /* filename */ comments to generated require()s in the output.
-      pathinfo: isEnvDevelopment,
-      // There will be one main bundle, and one file per asynchronous chunk.
-      // In development, it does not produce real files.
-      filename: isEnvProduction
-        ? 'static/js/[name].[contenthash:8].js'
-        : isEnvDevelopment && 'static/js/[name].js',
-      // There are also additional JS chunk files if you use code splitting.
-      // Please remember that Webpack 5, unlike Webpack 4, controls "splitChunks" via fileName, not chunkFilename - https://stackoverflow.com/questions/66077740/webpack-5-output-chunkfilename-not-working
-      chunkFilename: isEnvProduction
-        ? 'static/js/[name].[contenthash:8].chunk.js'
-        : isEnvDevelopment && 'static/js/[name].chunk.js',
       // webpack uses `publicPath` to determine where the app is being served from.
       // It requires a trailing slash, or the file assets will get an incorrect path.
       // We inferred the "public path" (such as / or /my-project) from homepage.
       publicPath: paths.publicUrlOrPath,
-      // Point sourcemap entries to original disk location (format as URL on Windows)
-      devtoolModuleFilenameTemplate: isEnvProduction
-        ? (info) =>
-            path
-              .relative(paths.appSrc, info.absoluteResourcePath)
-              .replace(/\\/g, '/')
-        : isEnvDevelopment &&
-          ((info) =>
-            path.resolve(info.absoluteResourcePath).replace(/\\/g, '/')),
-    },
-    optimization: {
-      minimize: isEnvProduction,
-      minimizer: [
-        // This is only used in production mode
-        new TerserPlugin({
-          terserOptions: {
-            parse: {
-              // We want terser to parse ecma 8 code. However, we don't want it
-              // to apply any minification steps that turns valid ecma 5 code
-              // into invalid ecma 5 code. This is why the 'compress' and 'output'
-              // sections only apply transformations that are ecma 5 safe
-              // https://github.com/facebook/create-react-app/pull/4234
-              ecma: 8,
-            },
-            compress: {
-              ecma: 5,
-              warnings: false,
-              // Disabled because of an issue with Uglify breaking seemingly valid code:
-              // https://github.com/facebook/create-react-app/issues/2376
-              // Pending further investigation:
-              // https://github.com/mishoo/UglifyJS2/issues/2011
-              comparisons: false,
-              // Disabled because of an issue with Terser breaking valid code:
-              // https://github.com/facebook/create-react-app/issues/5250
-              // Pending further investigation:
-              // https://github.com/terser-js/terser/issues/120
-              inline: 2,
-            },
-            mangle: {
-              safari10: true,
-            },
-            // Added for profiling in devtools
-            keep_classnames: isEnvProductionProfile,
-            keep_fnames: isEnvProductionProfile,
-            output: {
-              ecma: 5,
-              comments: false,
-              // Turned on because emoji and regex is not minified properly using default
-              // https://github.com/facebook/create-react-app/issues/2488
-              ascii_only: true,
-            },
-          },
-        }),
-        new CssMinimizerPlugin(),
-      ],
-      // Automatically split vendor and commons
-      // https://twitter.com/wSokra/status/969633336732905474
-      // https://medium.com/webpack/webpack-4-code-splitting-chunk-graph-and-the-splitchunks-optimization-be739a861366
-      splitChunks: isApp
-        ? {
-            chunks: 'all',
-          }
-        : undefined,
-      // Keep the runtime chunk separated to enable long term caching
-      // https://twitter.com/wSokra/status/969679223278505985
-      // https://github.com/facebook/create-react-app/issues/5358
-      runtimeChunk: isApp
-        ? {
-            name: (entrypoint) => `runtime-${entrypoint.name}`,
-          }
-        : undefined,
     },
     resolve: {
       // This allows you to set a fallback for where webpack should look for modules.
@@ -577,176 +513,92 @@ module.exports = function (webpackEnv) {
         },
       ],
     },
-    plugins: [
-      // Generates an `index.html` file with the <script> injected.
-      isApp
-        ? new HtmlWebpackPlugin(
-            Object.assign(
-              {},
-              {
-                inject: true,
-                template: paths.appHtml,
-              },
-              isEnvProduction
-                ? {
-                    minify: {
-                      removeComments: true,
-                      collapseWhitespace: true,
-                      removeRedundantAttributes: true,
-                      useShortDoctype: true,
-                      removeEmptyAttributes: true,
-                      removeStyleLinkTypeAttributes: true,
-                      keepClosingSlash: true,
-                      minifyJS: true,
-                      minifyCSS: true,
-                      minifyURLs: true,
-                    },
-                  }
-                : undefined,
-            ),
-          )
-        : isEsmViewDevelopment
-        ? // We need to provide a synthetic index.html in case we're starting a ESM view
-          new HtmlWebpackPlugin(
-            Object.assign(
-              {},
-              {
-                inject: true,
-                templateContent: `
-                <html>
-                  <body>
-                    <div id="root"></div>
-                  </body>
-                </html>
-                `,
-                scriptLoading: 'module',
-              },
-            ),
-          )
-        : false,
-      // Inlines the webpack runtime script. This script is too small to warrant
-      // a network request.
-      // https://github.com/facebook/create-react-app/issues/5358
-      isApp &&
-        isEnvProduction &&
-        shouldInlineRuntimeChunk &&
-        new InlineChunkHtmlPlugin(HtmlWebpackPlugin, [/runtime-.+[.]js/]),
-      // Makes some environment variables available in index.html.
-      // The public URL is available as %PUBLIC_URL% in index.html, e.g.:
-      // <link rel="icon" href="%PUBLIC_URL%/favicon.ico">
-      // It will be an empty string unless you specify "homepage"
-      // in `package.json`, in which case it will be the pathname of that URL.
-      isApp && new InterpolateHtmlPlugin(HtmlWebpackPlugin, env.raw),
-      // This gives some necessary context to module not found errors, such as
-      // the requesting resource.
-      new ModuleNotFoundPlugin(paths.appPath),
-      // Makes some environment variables available to the JS code, for example:
-      // if (process.env.NODE_ENV === 'production') { ... }. See `./env.js`.
-      // It is absolutely essential that NODE_ENV is set to production
-      // during a production build.
-      // Otherwise React will be compiled in the very slow development mode.
-      new webpack.DefinePlugin(env.stringified),
-      // Watcher doesn't work well if you mistype casing in a path so we use
-      // a plugin that prints an error when you attempt to do this.
-      // See https://github.com/facebook/create-react-app/issues/240
-      isEnvDevelopment && new CaseSensitivePathsPlugin(),
-      // If you require a missing module and then `npm install` it, you still have
-      // to restart the development server for webpack to discover it. This plugin
-      // makes the discovery automatic so you don't have to restart.
-      // See https://github.com/facebook/create-react-app/issues/186
-      isEnvDevelopment &&
-        new WatchMissingNodeModulesPlugin(paths.appNodeModules),
-      isEnvProduction &&
-        new MiniCssExtractPlugin({
-          // Options similar to the same options in webpackOptions.output
-          // both options are optional
-          filename: 'static/css/[name].[contenthash:8].css',
-          chunkFilename: 'static/css/[name].[contenthash:8].chunk.css',
-        }),
-      // Generate an asset manifest file with the following content:
-      // - "files" key: Mapping of all asset filenames to their corresponding
-      //   output file so that tools can pick it up without having to parse
-      //   `index.html`
-      // - "entrypoints" key: Array of files which are included in `index.html`,
-      //   can be used to reconstruct the HTML if necessary
-      new WebpackManifestPlugin({
-        fileName: 'asset-manifest.json',
-        publicPath: paths.publicUrlOrPath,
-        generate: (seed, files, entrypoints) => {
-          const manifestFiles = files.reduce((manifest, file) => {
-            manifest[file.name] = file.path;
-            return manifest;
-          }, seed);
-          const entrypointFiles = entrypoints.main.filter(
-            (fileName) => !fileName.endsWith('.map'),
-          );
-
-          return {
-            files: manifestFiles,
-            entrypoints: entrypointFiles,
-          };
-        },
-      }),
-      // Moment.js is an extremely popular library that bundles large locale files
-      // by default due to how webpack interprets its code. This is a practical
-      // solution that requires the user to opt into importing specific locales.
-      // https://github.com/jmblog/how-to-optimize-momentjs-with-webpack
-      // You can remove this if you don't use Moment.js:
-      new webpack.IgnorePlugin({
-        resourceRegExp: /^\.\/locale$/,
-        contextRegExp: /moment$/,
-      }),
-      // TypeScript type checking turned off for CI envs
-      // https://github.com/jpmorganchase/modular/issues/605
-      useTypeScript &&
-        !isCI &&
-        new ForkTsCheckerWebpackPlugin({
-          async: isEnvDevelopment,
-          typescript: {
-            async: isEnvDevelopment,
-            typescriptPath: resolve.sync('typescript', {
-              basedir: paths.appNodeModules,
-            }),
-            configOverwrite: {
-              compilerOptions: {
-                sourceMap: isEnvProduction
-                  ? shouldUseSourceMap
-                  : isEnvDevelopment,
-                skipLibCheck: true,
-                inlineSourceMap: false,
-                declarationMap: false,
-                noEmit: true,
-              },
+    optimization: {
+      minimizer: [
+        // This is only used in production mode
+        new TerserPlugin({
+          terserOptions: {
+            parse: {
+              // We want terser to parse ecma 8 code. However, we don't want it
+              // to apply any minification steps that turns valid ecma 5 code
+              // into invalid ecma 5 code. This is why the 'compress' and 'output'
+              // sections only apply transformations that are ecma 5 safe
+              // https://github.com/facebook/create-react-app/pull/4234
+              ecma: 8,
             },
-            context: paths.appPath,
-            diagnosticOptions: {
-              syntactic: true,
-              semantic: true,
+            compress: {
+              ecma: 5,
+              warnings: false,
+              // Disabled because of an issue with Uglify breaking seemingly valid code:
+              // https://github.com/facebook/create-react-app/issues/2376
+              // Pending further investigation:
+              // https://github.com/mishoo/UglifyJS2/issues/2011
+              comparisons: false,
+              // Disabled because of an issue with Terser breaking valid code:
+              // https://github.com/facebook/create-react-app/issues/5250
+              // Pending further investigation:
+              // https://github.com/terser-js/terser/issues/120
+              inline: 2,
             },
-            mode: 'write-references',
-          },
-          issue: {
-            // This one is specifically to match during CI tests,
-            // as micromatch doesn't match
-            // '../cra-template-typescript/template/src/App.tsx'
-            // otherwise.
-            include: [
-              { file: '../**/src/**/*.{ts,tsx}' },
-              { file: '**/src/**/*.{ts,tsx}' },
-            ],
-            exclude: [
-              { file: '**/src/**/__tests__/**' },
-              { file: '**/src/**/?(*.){spec|test}.*' },
-              { file: '**/src/setupProxy.*' },
-              { file: '**/src/setupTests.*' },
-            ],
+            mangle: {
+              safari10: true,
+            },
+            // Added for profiling in devtools
+            keep_classnames: isEnvProductionProfile,
+            keep_fnames: isEnvProductionProfile,
+            output: {
+              ecma: 5,
+              comments: false,
+              // Turned on because emoji and regex is not minified properly using default
+              // https://github.com/facebook/create-react-app/issues/2488
+              ascii_only: true,
+            },
           },
         }),
-    ].filter(Boolean),
+        new CssMinimizerPlugin(),
+      ],
+    },
     // Turn off performance processing because we utilize
     // our own hints via the FileSizeReporter
     performance: false,
   };
+
+  const pluginConfig = createPluginConfig({
+    isApp,
+    isEnvProduction,
+    shouldUseSourceMap,
+    useTypeScript,
+  });
+
+  const webpackConfig = merge([
+    baseConfig,
+    isApp ? appConfig : esmViewConfig,
+    isEnvProduction ? productionConfig : developementConfig,
+    pluginConfig,
+  ]);
+
+  function rewriteExternals({ request }, callback) {
+    const parsedModule = parsePackageName(request);
+
+    // If the module is absolute and it is in the import map, we want to externalise it
+    if (
+      parsedModule &&
+      parsedModule.dependencyName &&
+      dependencyMap[parsedModule.dependencyName] &&
+      // If this is an absolute export of css we need to deal with it in the loader
+      !request.endsWith('.css')
+    ) {
+      const { dependencyName, submodule } = parsedModule;
+
+      const toRewrite = `${dependencyMap[dependencyName]}${
+        submodule ? `/${submodule}` : ''
+      }`;
+
+      return callback(null, toRewrite);
+    }
+    // Otherwise we just want to bundle it
+    return callback();
+  }
 
   // These dependencies are so widely used for us (JPM) that it makes sense to install
   // their webpack plugin when used.
